docs+engine: reconcile state drift per user decisions
- state.md: rename compassion → altruism (§Personality + §AgentData.traits)
  to match engine's hot_altruism from the state-port
- state.md + engine: Inventory.gold u32 → i64 (debt as signed negative);
  reconciled against Ability Plan 1's TransferGold i32
- schema_hash.rs fingerprint updated; baseline regenerated to
  3ac38a40271f... 210 tests still green, clippy clean.

+++ crates/engine/src/state/agent_types.rs
@@ -64,12 +64,13 @@ pub struct Membership {
 
 // ---- Task H: Inventory ---------------------------------------------------
 
-/// Portable commodity storage, one per agent. Minimal MVP: `gold` as unsigned
-/// (matches state.md §AgentData type modulo a later debt plan) plus an
-/// 8-commodity fixed slot table. Real item entities are a later plan.
+/// Portable commodity storage, one per agent. `gold` is signed (`i64`) so debt
+/// is representable as a negative balance — reconciled 2026-04-19 with state.md
+/// §AgentData and Ability Plan 1's `TransferGold` effect op. 8-commodity fixed
+/// slot table. Real item entities are a later plan.
 #[derive(Copy, Clone, Debug, Default, PartialEq, Eq)]
 pub struct Inventory {
-    pub gold:        u32,
+    pub gold:        i64,
     pub commodities: [u16; 8],
 }
